Implements time-of-stream synchronization.
In some cases, the input doesn't behave like a real-time stream and passes, at
once, a lot more data than it is possible to play. This happens with some
applications or when jumping the stream.

This code tracks the "time of stream" according to the configured audio rate
and sample size.  If there is more data than could be played in real time, the
packetizer waits.

The testcase now works perfectly, but the code wasn't yet tested live.