Refactor chunk player into multiple methods and two classes.
A lot was happening in this single class.

This is a preparation to include the output buffer size in the chunk
desired_time calculations. I'm trying to make Wavesync run on a single core
device which requires a particularly large output buffer (it even can't play
with mplayer/cvlc without hiccups). I hope that fixing this case can make it
more robust on other devices as well.

Also, fix sink_latency second/milisecond mistake - fixed by erwin8086 in
separate fork on Github.

Someone tried to replace pyaudio with something else - extracted pyaudio to
separate class so it should be simpler now.